prepare gql server for npm export
