+++ src/index.ts
@@ -1,0 +1,2 @@
+export { createGraphQLServer } from './http/graphqlServer.js';
+export type { GraphQLContext } from './context.js';
